fix: rename _utils to _utility.general_utils to avoid conflict between 1.10.0 and future releases
