+++ src/promptflow-azure/promptflow/azure/_entities/_flow.py
@@ -13,8 +13,8 @@ import pydash
 from promptflow._constants import FlowLanguage
 from promptflow._proxy import ProxyFactory
 from promptflow._sdk._constants import SERVICE_FLOW_TYPE_2_CLIENT_FLOW_TYPE, AzureFlowSource, FlowType
+from promptflow._sdk._utility.general_utils import PromptflowIgnoreFile, load_yaml, remove_empty_element_from_dict
 from promptflow._sdk._utility.signature_utils import update_signatures
-from promptflow._sdk._utils import PromptflowIgnoreFile, load_yaml, remove_empty_element_from_dict
 from promptflow._utils.flow_utils import dump_flow_dag, load_flow_dag, resolve_flow_path
 from promptflow._utils.logger_utils import LoggerFactory
 from promptflow.azure._ml import AdditionalIncludesMixin, Code
@@ -180,7 +180,7 @@ class Flow(AdditionalIncludesMixin):
         """Get all additional include configs.
         For flow, its additional include need to be read from dag with a helper function.
         """
-        from promptflow._sdk._utils.general_utils import _get_additional_includes
+        from promptflow._sdk._utility.general_utils import _get_additional_includes
 
         return _get_additional_includes(os.path.join(self.code, self.path))
