+++ src/promptflow-azure/promptflow/azure/_pf_client.py
@@ -10,7 +10,7 @@ from azure.core.credentials import TokenCredential
 
 from promptflow._sdk._constants import MAX_SHOW_DETAILS_RESULTS
 from promptflow._sdk._errors import RunOperationParameterError
-from promptflow._sdk._utils import generate_yaml_entry
+from promptflow._sdk._utility.general_utils import generate_yaml_entry
 from promptflow._sdk.entities import Run
 from promptflow._utils.user_agent_utils import ClientUserAgentUtil, setup_user_agent_to_operation_context
 from promptflow.azure._restclient.service_caller_factory import _FlowServiceCallerFactory
