+++ src/promptflow-azure/promptflow/azure/_storage/cosmosdb/summary.py
@@ -19,7 +19,7 @@ from promptflow._sdk._constants import (
     SPAN_EVENTS_NAME_PF_OUTPUT,
     TRACE_DEFAULT_COLLECTION,
 )
-from promptflow._sdk._utils import json_loads_parse_const_as_str
+from promptflow._sdk._utility.general_utils import json_loads_parse_const_as_str
 from promptflow._sdk.entities._trace import Span
 from promptflow.azure._storage.cosmosdb.cosmosdb_utils import safe_create_cosmosdb_item
