+++ src/promptflow-azure/promptflow/azure/operations/_run_operations.py
@@ -45,7 +45,12 @@ from promptflow._sdk._constants import (
 )
 from promptflow._sdk._errors import InvalidRunStatusError, RunNotFoundError, RunOperationParameterError
 from promptflow._sdk._telemetry import ActivityType, WorkspaceTelemetryMixin, monitor_operation
-from promptflow._sdk._utils import incremental_print, is_multi_container_enabled, is_remote_uri, print_red_error
+from promptflow._sdk._utility.general_utils import (
+    incremental_print,
+    is_multi_container_enabled,
+    is_remote_uri,
+    print_red_error,
+)
 from promptflow._sdk.entities import Run
 from promptflow._utils.async_utils import async_run_allowing_running_loop
 from promptflow._utils.logger_utils import get_cli_sdk_logger
