+++ src/promptflow-devkit/promptflow/_cli/_params.py
@@ -6,7 +6,7 @@ import argparse
 
 from promptflow._cli._completers._param_completers import run_name_completer
 from promptflow._sdk._constants import PROMPT_FLOW_DIR_NAME, PROMPT_FLOW_RUNS_DIR_NAME, CLIListOutputFormat, FlowType
-from promptflow._sdk._utils import load_input_data
+from promptflow._sdk._utility.general_utils import load_input_data
 
 # TODO: avoid azure dependency here
 MAX_LIST_CLI_RESULTS = 50
