+++ src/promptflow-devkit/promptflow/_cli/_pf/entry.py
@@ -32,7 +32,7 @@ from promptflow._cli._pf._trace import add_trace_parser, dispatch_trace_cmds  # 
 from promptflow._cli._pf._upgrade import add_upgrade_parser, upgrade_version  # noqa: E402
 from promptflow._cli._pf.help import show_privacy_statement, show_welcome_message  # noqa: E402
 from promptflow._cli._user_agent import USER_AGENT  # noqa: E402
-from promptflow._sdk._utils import print_pf_version, print_promptflow_version_dict_string  # noqa: E402
+from promptflow._sdk._utility.general_utils import print_pf_version, print_promptflow_version_dict_string  # noqa: E402
 from promptflow._utils.logger_utils import get_cli_sdk_logger  # noqa: E402
 from promptflow._utils.user_agent_utils import setup_user_agent_to_operation_context  # noqa: E402
