+++ src/promptflow-devkit/promptflow/_internal/__init__.py
@@ -50,7 +50,7 @@ from promptflow._proxy._csharp_executor_proxy import CSharpBaseExecutorProxy
 from promptflow._sdk._constants import LOCAL_MGMT_DB_PATH, CreatedByFieldName
 from promptflow._sdk._service.apis.collector import trace_collector
 from promptflow._sdk._tracing import process_otlp_trace_request
-from promptflow._sdk._utils.general_utils import resolve_flow_language
+from promptflow._sdk._utility.general_utils import resolve_flow_language
 from promptflow._sdk._version import VERSION
 from promptflow._utils.context_utils import _change_working_dir, inject_sys_path
 from promptflow._utils.credential_scrubber import CredentialScrubber
