+++ src/promptflow-devkit/promptflow/_sdk/_orchestrator/test_submitter.py
@@ -15,7 +15,7 @@ from promptflow._core._errors import NotSupported
 from promptflow._internal import ConnectionManager
 from promptflow._proxy import ProxyFactory
 from promptflow._sdk._constants import PROMPT_FLOW_DIR_NAME
-from promptflow._sdk._utils import get_flow_name, get_flow_path
+from promptflow._sdk._utility.general_utils import get_flow_name, get_flow_path
 from promptflow._sdk.entities._flows import Flow, FlowContext, Prompty
 from promptflow._sdk.operations._local_storage_operations import LoggerOperations
 from promptflow._utils.async_utils import async_run_allowing_running_loop
