+++ src/promptflow-devkit/promptflow/_sdk/_tracing.py
@@ -51,8 +51,11 @@ from promptflow._sdk._service.utils.utils import (
     is_port_in_use,
     is_run_from_built_binary,
 )
+from promptflow._sdk._utility.general_utils import (
+    add_executable_script_to_env_path,
+    extract_workspace_triad_from_trace_provider,
+)
 from promptflow._sdk._utility.tracing_utils import get_workspace_kind, parse_kv_from_pb_attribute, parse_protobuf_span
-from promptflow._sdk._utils import add_executable_script_to_env_path, extract_workspace_triad_from_trace_provider
 from promptflow._utils.logger_utils import get_cli_sdk_logger
 from promptflow._utils.thread_utils import ThreadWithContextVars
 from promptflow.tracing._integrations._openai_injector import inject_openai_api
