+++ src/promptflow-devkit/promptflow/_sdk/_utility/__init__.py
@@ -7,5 +7,5 @@
 # the directory as _utils, it won't be removed when installing promptflow < 1.8.0 with promptflow >= 1.8.0 installed,
 # and both the _utils directory and old _utils.py file will exist in the site-packages directory and cause import error.
 # So we need to rename the directory to _utility to avoid this issue.
-# TODO: move _utils.py into _utility and rename _utility to _utils after promptflow-runtime start  to give warning
+# TODO: rename _utility to _utils after promptflow-runtime start  to give warning
 #   on environment with lower version of promptflow
