+++ src/promptflow-devkit/promptflow/_sdk/_visualize_functions.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import Optional
 
 from promptflow._sdk._constants import VIS_HTML_TMPL
-from promptflow._sdk._utils import render_jinja_template
+from promptflow._sdk._utility.general_utils import render_jinja_template
 from promptflow.contracts._run_management import VisualizationRender
