+++ src/promptflow-devkit/promptflow/_sdk/entities/_yaml_translatable.py
@@ -5,7 +5,7 @@ import abc
 from typing import Dict, Optional
 
 from promptflow._sdk._constants import BASE_PATH_CONTEXT_KEY, CommonYamlFields
-from promptflow._sdk._utils import load_from_dict
+from promptflow._sdk._utility.general_utils import load_from_dict
 from promptflow._utils.yaml_utils import dump_yaml
