+++ src/promptflow-devkit/promptflow/_sdk/operations/_local_storage_operations.py
@@ -23,8 +23,7 @@ from promptflow._sdk._constants import (
     LocalStorageFilenames,
 )
 from promptflow._sdk._errors import BulkRunException, InvalidRunError
-from promptflow._sdk._utility.signature_utils import update_signatures
-from promptflow._sdk._utils import (
+from promptflow._sdk._utility.general_utils import (
     PromptflowIgnoreFile,
     generate_flow_tools_json,
     is_flex_run,
@@ -35,6 +34,7 @@ from promptflow._sdk._utils import (
     read_open,
     write_open,
 )
+from promptflow._sdk._utility.signature_utils import update_signatures
 from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._flows import Flow
 from promptflow._utils.exception_utils import PromptflowExceptionPresenter
