+++ src/promptflow-devkit/promptflow/_sdk/operations/_run_operations.py
@@ -22,7 +22,7 @@ from promptflow._sdk._constants import (
 from promptflow._sdk._errors import InvalidRunStatusError, RunExistsError, RunNotFoundError, RunOperationParameterError
 from promptflow._sdk._orm import RunInfo as ORMRun
 from promptflow._sdk._telemetry import ActivityType, TelemetryMixin, monitor_operation
-from promptflow._sdk._utils import incremental_print, print_red_error, safe_parse_object_list
+from promptflow._sdk._utility.general_utils import incremental_print, print_red_error, safe_parse_object_list
 from promptflow._sdk._visualize_functions import dump_html, generate_html_string
 from promptflow._sdk.entities import Run
 from promptflow._sdk.operations._local_storage_operations import LocalStorageOperations
