+++ src/promptflow-devkit/tests/sdk_cli_test/unittests/test_run.py
@@ -14,7 +14,7 @@ from promptflow._sdk._load_functions import load_flow, load_run
 from promptflow._sdk._orchestrator import RunSubmitter, overwrite_variant, variant_overwrite_context
 from promptflow._sdk._pf_client import PFClient
 from promptflow._sdk._run_functions import create_yaml_run
-from promptflow._sdk._utils import callable_to_entry_string
+from promptflow._sdk._utility.general_utils import callable_to_entry_string
 from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._flows import Flow
 from promptflow._sdk.operations._local_storage_operations import LocalStorageOperations
